mumble-bot/requirements.txt
pymumble
faster-whisper
piper-tts
webrtcvad-wheels
numpy
scipy
requests
redis
